Create PageObjectArgumentResolver.
Unfortunately it will not work with the way contexts are currently initialised, as mink session is prepared on before scenario or example events, and context is initialised on before suite. Preparing default mink session on before suite event fixes the problem, but in some cases page object would get a wrong session.