Add MinkExtension as extension requirement
Current documentation states that only Behat 3.0+ is
required for BehatPageObjectExtension to work.

By following the bootstrap tutorial in the docs, user
might be misled when launching Behat for the first time
as she will probably end up receiving an exception by telling
that a "non-existent parameter 'mink.parameter'" hasn't been
specified.

This commit should let users be aware of the explicit requirement
the extension has to have in place before using it.